Keccak: use size_t for function prototypes

+++ hash/src/asm/ARMv8A/KeccakP-1600-SnP.h
@@ -44,8 +44,8 @@
 #define KeccakP1600_StaticInitialize()
 void KeccakP1600_Initialize(void *state);
 void KeccakP1600_AddByte(void *state, unsigned char data, unsigned int offset);
-void KeccakP1600_AddBytes(void *state, const unsigned char *data,
-			  unsigned int offset, unsigned int length);
+void KeccakP1600_AddBytes(void *state, const unsigned char *data, size_t offset,
+			  size_t length);
 void KeccakP1600_OverwriteBytes(void *state, const unsigned char *data,
 				unsigned int offset, unsigned int length);
 void KeccakP1600_OverwriteWithZeroes(void *state, unsigned int byteCount);
